% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.ddmmmm.R
\name{convert.ddmmmm}
\alias{convert.ddmmmm}
\title{convert.ddmmmm}
\usage{
convert.ddmmmm(df = NULL, lat.field = "LAT", lon.field = "LON")
}
\arguments{
\item{lat.field}{the default is \code{"LAT"}. the name of the 
field holding latitude values (in DDMMMM)}

\item{lon.field}{the default is \code{"LON"}.  the name of the 
field holding longitude values (in DDMMMM)}

\item{\code{df}}{= dataframe with coordinate fields}
}
\description{
This function can convert coordinates formatted as DDMMMM to 
decimal degrees.  This script will ignore NA values, and will retain the sign 
of the input coord
}
\examples{
lats = c(430925,-430925,432519,432519,430954,NA)
longs = c(341249,474550,-474550,NA,695860,-625950)
test=as.data.frame(cbind(lats,longs))
colnames(test)=c("LATITUDEDDMMMM","LONGITUDEDDMMMM")
 convert.ddmmmm(test, lat.field = "LATITUDEDDMMMM", lon.field = "LONGITUDEDDMMMM")
}
\author{
Mike McMahon
}
