% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrimp.db.r
\name{shrimp.db}
\alias{shrimp.db}
\title{shrimp.db}
\usage{
shrimp.db(DS = "complete.redo", oracle.server = "PTRAN",
  oracle.username = NULL, oracle.password = NULL,
  datadirectory = file.path(getwd(), "data"))
}
\arguments{
\item{DS}{is the main switch that selects which data source to load or operate. 
Options for DS include: 'complete','survey','ComLogs'. Any of these arguements called as listed return the data object. To make the data file from scratch would require a 'XXXX.redo', where XXXX is the option listed above.}

\item{oracle.server}{This is the server}

\item{oracle.username}{This is the username}

\item{oracle.password}{This is the password}

\item{datadirectory}{This is where your data will be saved}
}
\value{
Data objects that contain the data for use in further analyses.
}
\description{
This function is the main workhorse to pull data from databases and some initial filtering of data used in shrimp stock assessments. Results are saved and can be reloaded using this function.
}
